新しいCコンパイラ対応のための変更です。結果としていろいろ問題が生じました。 1. stdarg 関係
比較関数が正しくうごかない理由を調べたところ、yylval 構造体における整数が32bit であることが原因
と思われた. 具体的には mkfnode に正しく 32bit 引数が渡らない(上位ビット?にゴミが入る)という症状
だった. これを long に変更したところとりあえず比較については問題が解消した.

2. ユーザ関数関係

簡単な関数を書いてみると引数が正しく渡らないので, 仮引数における変数の番号が,
やはり上位ビットにゴミが入るのが原因のようなのでこれを生成する makepvar() が unsigned long を
返すように変更した.

3. map 関係

map が正しく動かない場合が見つかり, 原因が I_MAP コマンドの第3引数が 0 の場合, やはりゴミが
入る場合があるようなので, 0L に変更した.

以上により, 少し大きめのプログラム (noro_pd.rr など)は動いているが, まだ問題はありそうなので
testing branch で調べる.

+++ asir2018/parse/pvar.c
@@ -130,7 +130,7 @@ void poppvs() {
 #define IS_PATTERN 3
 
 
-unsigned int makepvar(char *str)
+unsigned long makepvar(char *str)
 {
   int c,c1,created;
   char *buf;
